Move PauseObservable and S3ResumeToken out of internal
