+++ services-custom/s3-transfer-manager/src/test/java/software/amazon/awssdk/transfer/s3/internal/DefaultFileUploadTest.java
@@ -36,8 +36,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import software.amazon.awssdk.services.s3.internal.multipart.PausableUpload;
-import software.amazon.awssdk.services.s3.internal.multipart.PauseObservable;
-import software.amazon.awssdk.services.s3.internal.multipart.S3ResumeToken;
+import software.amazon.awssdk.services.s3.multipart.PauseObservable;
+import software.amazon.awssdk.services.s3.multipart.S3ResumeToken;
 import software.amazon.awssdk.services.s3.internal.multipart.UploadWithKnownContentLengthHelper;
 import software.amazon.awssdk.services.s3.model.PutObjectResponse;
 import software.amazon.awssdk.transfer.s3.internal.model.DefaultFileUpload;
