+++ services/s3/src/main/java/software/amazon/awssdk/services/s3/multipart/S3PauseResumeExecutionAttribute.java
@@ -13,15 +13,14 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.services.s3.internal.multipart;
+package software.amazon.awssdk.services.s3.multipart;
 
 import software.amazon.awssdk.annotations.SdkProtectedApi;
-import software.amazon.awssdk.services.s3.multipart.S3ResumeToken;
+import software.amazon.awssdk.core.interceptor.ExecutionAttribute;
+import software.amazon.awssdk.core.interceptor.SdkExecutionAttribute;
 
 @SdkProtectedApi
-public interface PausableUpload {
-
-    default S3ResumeToken pause() {
-        throw new UnsupportedOperationException();
-    }
+public final class S3PauseResumeExecutionAttribute extends SdkExecutionAttribute {
+    public static final ExecutionAttribute<S3ResumeToken> RESUME_TOKEN = new ExecutionAttribute<>("ResumeToken");
+    public static final ExecutionAttribute<PauseObservable> PAUSE_OBSERVABLE = new ExecutionAttribute<>("PauseObservable");
 }
